+++ java/src/test/java/com/rojakcoder/archly/PermissionTest.java
@@ -9,18 +9,19 @@ public class PermissionTest {
 	static Permission p = Permission.getSingleton();
 
 	@Test(priority = 3)
-	public void testRunner() {
+	public void testRunner() throws InterruptedException {
 		testIsAllowedDenied();
 		testAllow();
 		testDeny();
 		testRemove();
+		testRemoveByResourceRole();
 
 		int size = p.size();
 		Permission pm = Permission.getSingleton();
 		Assert.assertEquals(pm.size(), size);
 	}
 
-	public void testIsAllowedDenied() {
+	private void testIsAllowedDenied() {
 		String res1 = new String("RES-1");
 		String rol1 = new String("ROLE-1");
 		String nullres = null;
@@ -59,7 +60,7 @@ public class PermissionTest {
 
 	}
 
-	public void testAllow() {
+	private void testAllow() {
 		String res1 = "RES-1";
 		String rol1 = "ROLE-1";
 
@@ -160,7 +161,7 @@ public class PermissionTest {
 		Assert.assertFalse(p.isDenied(rol1, res1, Permission.Types.DELETE));
 	}
 
-	public void testDeny() {
+	private void testDeny() {
 		//use different identifiers
 		String res1 = "RES-A";
 		String rol1 = "ROLE-A";
@@ -275,7 +276,7 @@ public class PermissionTest {
 		Assert.assertTrue(p.isDenied(rol1, res1, Permission.Types.DELETE));
 	}
 
-	public void testRemove() {
+	private void testRemove() {
 		String res1 = new String("RES-1");
 		String rol1 = new String("ROLE-1");
 		String resna = new String("RES-NA");
@@ -355,6 +356,103 @@ public class PermissionTest {
 			thrown = true;
 		}
 		Assert.assertTrue(thrown); //exception when repeated
+	}
+
+	private void testRemoveByResourceRole() throws InterruptedException {
+		String resources[] = {
+			"Q1", "Q2", "Q3", "Q4"
+		};
+		String roles[] = {
+			"P1", "P2", "P3", "P4"
+		};
+		p.clear();
+
+		Assert.assertEquals(p.size(), 0);
+		//create mappings for each key pair
+		for (String res: resources) {
+			for (String rol: roles) {
+				p.allow(rol, res);
+			}
+		}
+		Assert.assertEquals(p.size(), 16); //4x4
+
+		//add ALL access
+		for (String res: resources) {
+			p.allow(null, res);
+		}
+		Assert.assertEquals(p.size(), 20); //4x4 + 4
+		for (String rol: roles) {
+			p.allow(rol, null);
+		}
+		Assert.assertEquals(p.size(), 24); //4x4 + 4 + 4
+
+		//remove all access on Q4
+		int removed = p.removeByResource(resources[3]);
+		Assert.assertEquals(p.size(), 19); //less all Q4
+		Assert.assertEquals(removed, 5);
+
+		//repeated removal should yield 0
+		removed = p.removeByResource(resources[3]);
+		Assert.assertEquals(p.size(), 19);
+		Assert.assertEquals(removed, 0); //none removed
+
+		//remove all access from P4
+		removed = p.removeByRole(roles[3]);
+		Assert.assertEquals(p.size(), 15); //less all P4 (Q4 already removed)
+		Assert.assertEquals(removed, 4);
+
+		//repeated removal should yield 0
+		removed = p.removeByRole(roles[3]);
+		Assert.assertEquals(p.size(), 15);
+		Assert.assertEquals(removed, 0); //none removed
+
+		//test coverage
+		//add many entries
+		for (int i = 0; i < 50; i++) {
+			p.allow("T" + i, "TA");
+			p.allow("TB", "T" + i);
+		}
+		//create thread 1 to call removeByResource
+		Thread t1 = new Thread(new Runnable() {
+			public void run() {
+				System.out.println(" > Removing by resource TA");
+				int rem = p.removeByResource("TA");
+				System.out.println("Actually removed resources: " + rem);
+				Assert.assertEquals(rem, 49); //T40 is removed by thread 3
+			}
+		});
+		//thread 2 to call removeByRole
+		Thread t2 = new Thread(new Runnable() {
+			public void run() {
+				System.out.println(" > Removing by role TB");
+				int rem = p.removeByRole("TB");
+				System.out.println("Actually removed roles: " + rem);
+				Assert.assertEquals(rem, 49); //T40 is removed by thread 3
+			}
+		});
+		//create thread 3 to remove specific resource
+		Thread t3 = new Thread(new Runnable() {
+			public void run() {
+				try {
+					p.remove("T40", "TA");
+				} catch (EntryNotFoundException e) {
+					System.out.println(" > Entry not found");
+				}
+				try {
+					p.remove("TB", "T40");
+				} catch (EntryNotFoundException e) {
+					System.out.println(" > Entry not found");
+				}
+			}
+		});
+		Permission.RUNTEST = true;
+		t1.start();
+		t2.start();
+		t3.start();
+		t1.join();
+		t2.join();
+		Permission.RUNTEST = false;
+		System.out.println(" > Done testing concurrent modifications");
 	}
 }